Refactor to modular package and drop device manager
