+++ piviewer/utils.py
@@ -9,7 +9,7 @@ import random
 import psutil
 from datetime import datetime
 
-from config import (
+from .config import (
     APP_VERSION,
     VIEWER_HOME,
     IMAGE_DIR,
@@ -22,9 +22,6 @@ def init_config():
     if not os.path.exists(CONFIG_PATH):
         default_cfg = {
             "theme": "dark",
-            "role": "main",
-            "main_ip": "",
-            "devices": [],
             # displays dictionary for multi-display logic
             "displays": {
                 "Display0": {
@@ -160,45 +157,4 @@ def count_files_in_folder(folder_path):
         if f.lower().endswith(valid_ext):
             cnt += 1
     return cnt
-
-################################
-# Remote device push/pull logic
-################################
-
-def get_remote_config(ip):
-    url = f"http://{ip}:8080/sync_config"
-    try:
-        r = requests.get(url, timeout=5)
-        if r.status_code == 200:
-            return r.json()
-    except Exception as e:
-        log_message(f"Error fetching remote config from {ip}: {e}")
-    return None
-
-def get_remote_monitors(ip):
-    url = f"http://{ip}:8080/list_monitors"
-    try:
-        r = requests.get(url, timeout=5)
-        if r.status_code == 200:
-            return r.json()
-    except Exception as e:
-        log_message(f"Error fetching remote monitors from {ip}: {e}")
-    return {}
-
-def push_displays_to_remote(ip, displays_obj):
-    url = f"http://{ip}:8080/update_config"
-    partial = {"displays": displays_obj}
-    try:
-        r = requests.post(url, json=partial, timeout=5)
-        if r.status_code == 200:
-            log_message(f"Pushed partial displays to {ip} successfully.")
-        else:
-            log_message(f"Push to {ip} failed with code {r.status_code}.")
-    except Exception as e:
-        log_message(f"Error pushing partial displays to {ip}: {e}")
-
-def pull_displays_from_remote(ip):
-    remote_cfg = get_remote_config(ip)
-    if not remote_cfg:
-        return None
-    return remote_cfg.get("displays", {})
+
