--- templates/device_manager.html
@@ -1,52 +1,0 @@
-{% extends "base.html" %}
-{% block title %}Manage Devices{% endblock %}
-{% block content %}
-
-<div class="page-section" style="max-width:1200px;">
-  <h2>Sub-Device Manager</h2>
-  <p>You are on the Main device. Manage sub-devices below.</p>
-
-  <form method="POST" style="padding:10px; border:1px solid var(--border-muted); border-radius:6px; margin-bottom:20px;">
-    <h3>Add Device</h3>
-    <label>Device Name:</label><br>
-    <input type="text" name="dev_name" required>
-    <br><br>
-    <label>Device IP (or hostname):</label><br>
-    <input type="text" name="dev_ip" required>
-    <br><br>
-    <button type="submit" name="action" value="add_device">Add Device</button>
-  </form>
-
-  <table>
-    <thead>
-      <tr>
-        <th>#</th>
-        <th>Device Name</th>
-        <th>IP</th>
-        <th>Actions</th>
-      </tr>
-    </thead>
-    <tbody>
-      {% for dev in cfg.devices|default([]) %}
-      <tr>
-        <td>{{ loop.index }}</td>
-        <td>{{ dev.name }}</td>
-        <td>{{ dev.ip }}</td>
-        <td>
-          <button type="submit" form="deviceActionForm" name="action" value="push_{{ loop.index0 }}">Push</button>
-          <button type="submit" form="deviceActionForm" name="action" value="pull_{{ loop.index0 }}">Pull</button>
-          <button type="submit" form="deviceActionForm" name="action" value="remove_{{ loop.index0 }}">Remove</button>
-          <form action="{{ url_for('main.remote_configure', dev_index=loop.index0) }}" method="GET" style="display:inline;">
-            <button type="submit">Configure</button>
-          </form>
-        </td>
-      </tr>
-      {% endfor %}
-    </tbody>
-  </table>
-
-  <form method="POST" id="deviceActionForm"></form>
-
-</div>
-
-{% endblock %}
