+++ templates/index.html
@@ -205,7 +205,7 @@
   <!-- Bottom status row as its own card -->
   <div class="card" style="margin-top:20px; text-align:center;">
     <strong>Status:</strong>
-    Spotify: {{ spotify_status }} | Weather: {{ weather_status }} | Subdevices: {{ subdevices_status }}
+    Spotify: {{ spotify_status }} | Weather: {{ weather_status }}
   </div>
 </div>
